+++ src/verified_contract/github.rs
@@ -3,7 +3,7 @@ use near_sdk::serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, BorshDeserialize, BorshSerialize)]
 #[serde(crate = "near_sdk::serde")]
-pub struct GithubData {
+pub struct Github {
     pub owner: String,
     pub repo: String,
     pub sha: String,
