+++ src/verified_contract/mod.rs
@@ -1,4 +1,4 @@
-pub mod github_data;
+pub mod github;
 pub mod vote;
 pub mod comment;
 
@@ -6,17 +6,17 @@ use std::collections::HashSet;
 use near_sdk::borsh::{self, BorshDeserialize, BorshSerialize};
 use near_sdk::serde::{Deserialize, Serialize};
 use vote::Vote;
-use github_data::GithubData;
+use github::Github;
 
 #[derive(Serialize, Deserialize, BorshDeserialize, BorshSerialize)]
 #[serde(crate = "near_sdk::serde")]
-pub struct ContractData {
+pub struct VerifiedContract {
     pub cid: String,
     pub lang: String,
     pub entry_point: String,
     pub code_hash: String,
     pub builder_image: String,
-    pub github: Option<GithubData>,
+    pub github: Option<Github>,
     pub votes: HashSet<Vote>,
     pub comments: Vec<u64>,
 }
