+++ src/verified_contract/vote.rs
